Extract NotificationSlot from NotificationActionsPreference
